Add block for building the OpenKIM library.
The Open Knowledgebase of Interatomic Models (openkim.org) is a library
of interatomic models that can be used by some simulation packages, including
LAMMPS and Asap.  An easybuild of the latter will be contributed shortly.

This easyblock compiles the library itself; then it downloads and compiles
all models currently available from the OpenKIM project.  Finally, it installs
everything and sets an environment variable pointing to the models.