Adding support for OPTARCH_GENERIC for GROMACS
This PR allows the use of EASYBUILD_OPTARCH=GENERIC for GROMACS. It sets
the GMX_SIMD to 'SSE2' if the underlying software is a X86_64. This is a
good approximation since SSE2 were introduced in 2001 for intel CPUs and
in 2003 for AMD ones. This build would only break if one is compiling
for a >18 year old intel or >15 years old AMD.

For other architectures it sets the GMX_SIMD to 'None'