If ebpythonprefixes=True, install sitecustomize.py.
This is put into $EBROOTPYTHON/easybuild/python, which is itself
added to PYTHONPATH, and listens to $EBPYTHONPREFIXES as a list
of prefixes to which lib/pythonx.y/site-packages is added for
every entry, where x.y depends on the running Python version.

$EBROOTPYTHON/easybuild/python needs to be put in PYTHONPATH
to make sure this directory is used even inside virtual environments.