[CMakeMake] Default CMAKE_BUILD_TYPE based on toolchainopts[debug]
If the build_type is not set explicitely, check toolchainopts[debug] and
use either 'Debug' or 'Release'.
The influence of 'Release' is rather minor as optarch is always used
which is probably set to some optimization level already and hence
overwrites the CMake-set optimization level.
It still is a good and reasonable default similar to e.g. ConfigureMake