[PDT/TAU] Generically find the arch folder and add top-level bin/lib symlinks
The arch folder is not $(uname -m), but determined by some custom if-chain
So short of parsing the configure output we simply symlink the bin/lib folder
of any folder containing those into the top-level folder.
This is what Spack does too so it is verified already

This fixes build on e.g. Power9