More flexible default_cuda_version
Default is now None, to be overwritten either via the easyconfig or the command line (--try-amend=default_cuda_version='11.0'). If it is not overwritten manually, it is tried to be guessed by requesting the version of a (potential) CUDA module. If that also fails, EB is error'd out with a descriptive message.