Add an easyblock for FSL installation via conda
FSL is now installed via a Conda environment. The default Conda easyblock cannot easily be used because:
a) FSL requires that certain environment variables are defined, e.g. FSLOUTPUTTYPE
so these would need to be added as an module file extra
b) The generic/conda block automatically adds various bin/library folders to system search paths, this is usually NOT what you want to do with FSL
c) If the variable FSL_CREATE_WRAPPER_SCRIPTS is set to '1' then wrapper scripts for FSL tools
will be created in FSLDIR/share/fsl/bin avoiding the polution of the user's PATHs
with conda provided bin/libraries.
EB_FSLCONDA creates a clean module file, ensures that these wrappers are created and that the minimum required environment variables are set in the generated module.