Use build dir for building TensorFlow pip package
build_pip_package (executed after Bazel finishes) uses a temporary
folder which gets large (2 GB for TF 2.9).
It uses `$TMPDIR` (via mktemp) which we set to the temporary folder
created by Easybuild in PythonPackage but that might have only enough
space for logs etc.
As this folder is part of the build put it inside the build directory.