[bugfix] Use correct temporary directory for python installation
There are 2 problems with the way the `pythonpath` instance field is defined:

- it uses the log directory to store a build artifact, and
- the functions that use `pythonpath` do not accept a full path.

We cannot use the `build_path()` function instead of `log_path()` as the
function provides a full path as well. Thus we opted to hard-code a
relative path with respect to the build directory.

Issue: https://github.com/easybuilders/easybuild-easyconfigs/issues/21078