disable the version check of pip in Python packages
When invoking `pip` it will connect to PyPI to check if a new version is
available. Not only do we not care about that but we also don't want to
connect to anything external which slows down the build at best or fails at worst.

Set the environment variable to disable this in the PythonPackage
constructor such that it works for any easyblock deriving from it.