make sure user packages are not used in sanity check of PythonBundle
Some PythonBundle ECs (e.g. IPython) use custom sanity checks using
python. That might pick up python packages installed to the users `$HOME`
which can yield wrong results.
We already take care of this for the extensions sanity check in PythonPackage and
for installing extensions in PythonBundle.
However for the sanity check of PythonBundle itself the variable is not
defined (even if it was defined during the build) as the environment
gets reset when loading the module used for the sanity check.
Hence use the same approach used in PythonPackage and set
`$PYTHONNOUSERSITE` after loading the module.