GCC: Fix NVPTX mapping error when with no CC
When no compute capability is set but NVPTX is enabled, trying to figure
out the NVPTX architecture fails with the error:

```
File "easybuild/easyblocks/g/gcc.py", line 431, in map_nvptx_capability
  return sorted_gcc_cc[0]
IndexError: list index out of range
```

The error occurs because of an insufficient check for an unset CUDA
compute capability. This commit changes the checked conditions, so that
empty lists are also correctly handled.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>