Enhance generic Bundle EasyBlock to transfer module requirements of components
This commit implements the `make_module_req_guess` method for the
generic Bundle EasyBlock. With this, all the requirements of the
components in a bundle are transferred correctly to the final module.
Previously, this could lead to missing environment variables, letting
the build succeed but still resulting in a broken module, for example
because `PATH` is not set.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>