Enhance AOCC EasyBlock to use config files for GCC toolchain when required
With AOCC 4.2.0 and 5.0.0, it was observed that wrappers for `clang++` and
`flang` will cause compile errors since some libraries are not linked anymore.
To work around this, create config files for `clang` and `clang++` instead.
Since `flang` doesn't understand config files yet, keep the old behavior.

Also add a sanity check to ensure that the correct toolchain is selected.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>
Co-authored-by: crivella <davide.grassano@epfl.ch>
Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>