score_p: Pass dependencies to EasyBlock based on EasyConfig
Instead of going through the defined dependencies in EasyBlock and
searching if they are loaded, go through the explicit dependencies in
the EasyConfig and pass them instead. This reduces the amount of
unnecessary passed configure options.

However, this now requires having 'binutils' as an explicit dependency
for Score-P, since it requires having the paths for libbfd.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>