AOCC: Use bash for compiler wrapper
AOCC 4.2.0 and lower (and 'flang' for AOCC 5.0.0) use compiler wrappers
to include the GCC toolchain in the compiler command. The original
compilers are renamed, and a shell script is created running 'exec'
which passes the original compiler name as an argument.

This however can fail if '/bin/sh' is used. On Ubuntu 22.04 LTS, using
'exec -a' causes the following error message, which then causes the
sanity check to fail:

```
$ exec -a "echo" echo "Hello World"
sh: 1: exec: -a: not found
```

To work around this, use '/bin/bash' instead.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>