impi: Add option to rebuild Fortran 2008 bindings
The Fortran 2008 bindings shiped with Intel MPI can cause
compilation failures when using `CLASS(*)` arguments.
This bug is known, and a solution for this given in the following
thread:

https://community.intel.com/t5/Intel-MPI-Library/MPI-f08-with-polymorphic-argument-CLASS/m-p/1590421

However, this bug is not yet included in Intel MPI releases to
preserve backwards compatibility. Instead, one needs to
rebuild the bindings themselves and replace them.
Add an option for this, and leave it disabled by default.
EasyConfigs can then decide if they want to rebuild these modules,
or preserve the backwards compatibility.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>