LLVM: Ignore failing nvptx64 tests without ptxas in PATH
LLVM can be built without needing CUDA, but tests will fail in that
case. To work around this issue, and avoid the need to load CUDA just to
test things, ignore these test failures when 'ptxas' is not present.

A similar workaround is likely needed for AMD as well, but needs to be
figured out first.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>