Don't modify self.cfg['configopts'] in CMakeMake easyblock
We use multi_deps a lot, e.g. in
https://github.com/ComputeCanada/easybuild-easyconfigs/blob/computecanada-main/easybuild/easyconfigs/r/RDKit/RDKit-2024.03.5-GCC-12.3.0.eb
the problem was now that CMakeMake automatically adds -DPython_EXECUTABLE
and -DPython3_EXECUTABLE pointing to e.g. first python 3.10 but in
the second iteration the logic keeps it that way in configopts, and it's wrong
since there it should be pointing to 3.11 instead.

We can avoid this by *not* modifying configopts, so it's reconstructed
from scratch in every iteration. Another way would be to use
$EBROOTPYTHON and let the shell expand the path, instead of
get_software_root().