LLVM: Use dlopen for libomptarget / offload
Avoid linking in a system CUDA / ROCm by mistake by forcing LLVM
to use dlopen the libraries instead. This also removes the need
to pass in these libraries as dependencies when building LLVM
altogether, except if testing of the built libraries is wanted.

This should only really matter for systems where a system ROCm
or CUDA exists.