lammps sanity checks: remove LD_LIBRARY_PATH setting
With the `foss-2024a` toolchain (CUDA 12.6+), prepending `LD_LIBRARY_PATH` with `LIBRARY_PATH` causes LAMMPS to crash, likely because the dynamic linker gets pointed to the CUDA stubs libs instead of the system libs.
Strangely, this is not a problem for older toolchains.
At the same time, it appears that setting `LD_LIBRARY_PATH` has no effect on the functionality of the code (it seems that Python is able to retrieve `liblammps.so` without it), so I propose we just get rid of it, unless there are actual scenarios where it causes problems.