Fix condition to ignore failed LLVM GPU tests
The tests should always be ignored if `ptxas` or the ROCm dependency is
not present independently of whether the corresponding target is in the
build targets.
At worst this will exclude too many tests but it seems the condition to
run those tests is independent of whether this specific target is build.