Handle crates in subfolders
The Cargo.toml file might contain:
```
[workspace]
members = ["components/salsa-macro-rules", "components/salsa-macros"]
```

This means we need to move the 2 subfolders of components and NOT the
components folder into the vendor root-directory so they are found.
Hence use `basename` for determining the target path which is a no-op
when no folder is present.