Error out when only build-options are given for PIP based Python package installation
When no separate build command is used to build Python packages,
e.g. when using pip, the build options will be ignored.
This may be surprising as it is different to any other build.

Error out in this case.

Do not error when build- and install-options are set as that might be
intentional to support run-time choices e.g. by easyblocks.
Also include a clear hint to pip if it is used.