Avoid PyTorch configure warnings/issues by explicitly setting (more) dependency options
Most of the options have a True/False value which we should set to
False/0 when we don't have/use that dependency.
This ensures that a) no system lib will be found and b) no warning will
be shown.

Also update the list with options added or removed until PyTorch 2.7