Symlink NCCL library when added as a build dependency
As PyTorch is sensitive to specific NCCL versions one approach is to use
it as a build dependency only and add an rpath to it after copying it
into a (non-standard) folder inside the PyTorch module.
This is similar to the PyPI package that depends on various
nvidia-packages and adds relative rpaths to ensure they are used when
loading the torch package/libraries.