[BUGFIX] Ensure that the gnupg-bundle supports tests for the gpgme component
The gnupg-bundle is using a UNIX socket path in its tests, specifically for the
gpgme component. However, UNIX socket paths are limited to 108 characters in
length. This easyblock ensures that the build directory is located in a path
sufficiently small so that the UNIX sockets for the tests can be created.

Source: https://gitlab.archlinux.org/archlinux/archlinux-wsl/-/issues/10