use native PLUMED by default with GROMACS 2026+
GROMACS 2025 was released on Feb 11, 2025. Changing EB's default behavior
of whether to apply patches or use GROMACS' native PLUMED support would
be surprising to users.

Changing the default with the yet unreleased GROMACS 2026 is safer.

In any case PLUMED patches for GROMACS 2025 and newer are based on
GROMACS' native support, which means that some PLUMED features that were
available with older GROMACS releases are not supported until GROMACS'
MDModule and the plugin mechansims have caught up again.

See also discussion in easybuilders/easybuild-easyblocks#3984